SAP visualizer: split for use in wasm
The binary is now only for native, and the egui ui is moved to the lib. This means when we want to add it to a wasm environment, we should be able to do so.
